+++ R/auto_fselector.R
@@ -1,7 +1,9 @@
-#' @title Syntactic Sugar for Automatic Feature Selection
+#' @title Function for Automatic Feature Selection
 #'
-#' @description
-#' Function to create an [AutoFSelector] object.
+#' @inherit AutoFSelector description
+#' @inheritSection AutoFSelector Resources
+#' @inherit AutoFSelector details
+#' @inheritSection AutoFSelector Nested Resampling
 #'
 #' @param method (`character(1)` | [FSelector])\cr
 #'  Key to retrieve fselector from [mlr_fselectors] dictionary or [FSelector] object.
@@ -12,11 +14,16 @@
 #' @param ... (named `list()`)\cr
 #'  Named arguments to be set as parameters of the fselector.
 #'
-#' @return [AutoFSelector]
+#' @return [AutoFSelector].
 #'
 #' @template param_learner
 #' @template param_resampling
 #' @template param_measure
+#' @template param_terminator
+#' @template param_store_fselect_instance
+#' @template param_store_benchmark_result
+#' @template param_store_models
+#' @template param_check_values
 #'
 #' @export
 #' @examples
@@ -28,14 +35,14 @@
 #'   term_evals = 4)
 #'
 #' at$train(tsk("pima"))
-auto_fselector = function(method, learner, resampling, measure, term_evals = NULL, term_time = NULL, ...) {
+auto_fselector = function(method, learner, resampling, measure = NULL, term_evals = NULL, term_time = NULL, terminator = NULL, store_fselect_instance = TRUE, store_benchmark_result = TRUE, store_models = FALSE, check_values = FALSE, ...) {
   fselector = if (is.character(method)) {
     assert_choice(method, mlr_fselectors$keys())
     fs(method, ...)
   } else {
     assert_fselector(method)
   }
-  terminator = terminator_selection(term_evals, term_time)
+  terminator = terminator %??% terminator_selection(term_evals, term_time)
 
-  AutoFSelector$new(learner, resampling, measure, terminator, fselector)
+  AutoFSelector$new(learner = learner, resampling = resampling, measure = measure, terminator = terminator, fselector = fselector, store_fselect_instance = store_fselect_instance, store_benchmark_result = store_benchmark_result, store_models = store_models, check_values = check_values)
 }