+++ man-roxygen/param_resampling.R
@@ -1,3 +1,4 @@
 #' @param resampling ([mlr3::Resampling])\cr
-#' Uninstantiated resamplings are instantiated during construction
-#' so that all configurations are evaluated on the same data splits.
+#'   Resampling that is used to evaluated the performance of the feature subsets.
+#'   Uninstantiated resamplings are instantiated during construction so that all feature subsets are evaluated on the same data splits.
+#'   Already instantiated resamplings are kept unchanged.